Incorporate custom beacon types
Add the ability to specify a custom beacon type using a parser layout
string and either a manufacturer ID or a service UUID.  Also included
are 3 default beacon types: AltBeacon, Eddystone UID, and Eddystone URL.

To register a beacon advertisement (i.e., broadcast a beacon
advertisement), you need to specify a type along with the ids and power.
The parser layout of the beacon type is then used to build the
advertisement data with the ids and power.