Rename 'parserLayout' and 'BeaconParser'
to 'beaconLayout' and 'BeaconLayout' because 'parser' is kind of
confusing.  This library doesn't scan and parse advertisements, it only
broadcasts them.  I was originally mimicking the names from the
ScanBeacon gem, but then I realized the Android Beacon Library calls it
a 'beaconLayout' string, not a 'parserLayout' string and I liked that
better.